added delete buttons

+++ app/Http/Controllers/CustomerController.php
@@ -110,7 +110,13 @@ class CustomerController extends Controller
      */
     public function destroy(Customer $customer)
     {
-        $customer->delete();
+        // $customer->delete();
+        // return response()->json(null, 204);
+        if ($customer) {
+            $customer->delete();
+        } else {
+            return response()->json(array(), 200);
+        }
         return response()->json(null, 204);
     }
